[FEATURE] 예약 관련 API 구현 #7
why
예약 데이터를 create,read,delete하기위한 api가 필요하다
what
로그인한 사용자에 대해서 예약데이터에 관한 조회가 가능하도록 처리
예약 데이터를 POST할때는 csrf 모듈을 이용하여 protection 미들웨어를 거치도록 처리

+++ server/database/migrations/20191015080934-create-reservation.js
@@ -7,7 +7,7 @@ module.exports = {
       type: Sequelize.BIGINT,
     },
     booker_id: {
-      type: Sequelize.INTEGER,
+      type: Sequelize.BIGINT,
       allowNull: false,
       foreignKey: true,
       references: {
@@ -16,7 +16,7 @@ module.exports = {
       },
     },
     room_id: {
-      type: Sequelize.INTEGER,
+      type: Sequelize.BIGINT,
       allowNull: false,
       foreignKey: true,
       references: {
